IBX-6276: Added removing content type draft when clicking cancel button
