+++ src/lib/Menu/Admin/ContentType/ContentTypeEditRightSidebarBuilder.php
@@ -6,89 +6,23 @@
  */
 namespace EzSystems\EzPlatformAdminUi\Menu\Admin\ContentType;
 
-use eZ\Publish\API\Repository\Exceptions as ApiExceptions;
-use EzSystems\EzPlatformAdminUi\Menu\AbstractBuilder;
 use EzSystems\EzPlatformAdminUi\Menu\Event\ConfigureMenuEvent;
-use EzSystems\EzPlatformAdminUi\Menu\MenuItemFactory;
 use JMS\TranslationBundle\Model\Message;
-use JMS\TranslationBundle\Translation\TranslationContainerInterface;
-use Knp\Menu\ItemInterface;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * KnpMenuBundle Menu Builder service implementation for AdminUI Section Edit contextual sidebar menu.
  *
  * @see https://symfony.com/doc/current/bundles/KnpMenuBundle/menu_builder_service.html
  */
-class ContentTypeEditRightSidebarBuilder extends AbstractBuilder implements TranslationContainerInterface
+class ContentTypeEditRightSidebarBuilder extends AbstractContentTypeRightSidebarBuilder
 {
     /* Menu items */
     const ITEM__SAVE = 'content_type_edit__sidebar_right__save';
     const ITEM__CANCEL = 'content_type_edit__sidebar_right__cancel';
 
-    /** @var \Symfony\Contracts\Translation\TranslatorInterface */
-    private $translator;
-
-    public function __construct(
-        MenuItemFactory $factory,
-        EventDispatcherInterface $eventDispatcher,
-        TranslatorInterface $translator
-    ) {
-        parent::__construct($factory, $eventDispatcher);
-
-        $this->translator = $translator;
-    }
-
-    /**
-     * @return string
-     */
     protected function getConfigureEventName(): string
     {
         return ConfigureMenuEvent::CONTENT_TYPE_EDIT_SIDEBAR_RIGHT;
-    }
-
-    /**
-     * @param array $options
-     *
-     * @return \Knp\Menu\ItemInterface
-     *
-     * @throws \InvalidArgumentException
-     * @throws ApiExceptions\BadStateException
-     * @throws \InvalidArgumentException
-     */
-    public function createStructure(array $options): ItemInterface
-    {
-        /** @var \Symfony\Component\Form\FormView $contentTypeEditFormView */
-        $contentTypeEditFormView = $options['form_view'];
-
-        /** @var \Knp\Menu\ItemInterface|\Knp\Menu\ItemInterface[] $menu */
-        $menu = $this->factory->createItem('root');
-
-        $menu->setChildren([
-            self::ITEM__SAVE => $this->createMenuItem(
-                self::ITEM__SAVE,
-                [
-                    'attributes' => [
-                        'class' => 'btn--trigger',
-                        'data-click' => sprintf('#%s', $contentTypeEditFormView['publishContentType']->vars['id']),
-                    ],
-                    'extras' => ['icon' => 'save'],
-                ]
-            ),
-            self::ITEM__CANCEL => $this->createMenuItem(
-                self::ITEM__CANCEL,
-                [
-                    'attributes' => [
-                        'class' => 'btn--trigger',
-                        'data-click' => sprintf('#%s', $contentTypeEditFormView['removeDraft']->vars['id']),
-                    ],
-                    'extras' => ['icon' => 'circle-close'],
-                ]
-            ),
-        ]);
-
-        return $menu;
     }
 
     /**
@@ -100,5 +34,15 @@ class ContentTypeEditRightSidebarBuilder extends AbstractBuilder implements Tran
             (new Message(self::ITEM__SAVE, 'menu'))->setDesc('Save'),
             (new Message(self::ITEM__CANCEL, 'menu'))->setDesc('Discard changes'),
         ];
+    }
+
+    public function getItemSaveIdentifier(): string
+    {
+        return self::ITEM__SAVE;
+    }
+
+    public function getItemCancelIdentifier(): string
+    {
+        return self::ITEM__CANCEL;
     }
 }
